modifications for functions
