+++ structures/function_variable.py
@@ -1,0 +1,32 @@
+
+
+class Variable:
+    def __init__(self, name, var_type):
+        self.__name = name
+        self.__type = var_type
+        self.__value = None
+
+    def name(self):
+        return self.__name
+
+    def var_type(self):
+        return self.__type
+
+    def value(self):
+        return self.__value
+
+
+class Function:
+    def __init__(self, name, func_type, parameters=[]):
+        self.__name = name
+        self.__type = func_type
+        self.__params = parameters
+
+    def name(self):
+        return self.__name
+
+    def return_type(self):
+        return self.__type
+
+    def params(self):
+        return self.__params
